Implemented core files deletion permissions check

+++ view/adminhtml/web/js/translation/columns.js
@@ -34,6 +34,7 @@ define(
                     {title: __('Updated'), field: 'file_update_time', sorter: 'string', visible: false},
                     {title: __('Rows'), field: 'rows_count', sorter: 'number', width: 85},
                     {title: __('Errors'), field: 'errors', sorter: 'number', width: 100},
+                    {title: __('Is core'), field: 'is_core', sorter: 'number', width: 100, visible: false},
                     {title: __('Type'), field: 'file_type', sorter: 'string', width: 100},
                     {title: __('Group'), field: 'file_group', sorter: 'string', width: 100, visible: false},
                     {title: __('Locale'), field: 'file_locale', sorter: 'string', width: 100}
@@ -57,6 +58,12 @@ define(
                                 // Get the row data
                                 var rowData = row.getData();
     
+                                // Check core file deletion
+                                if (com.options.settings.allow_core_files_deletion == '0' && rowData.is_core) {
+                                    alert('not allowed');
+                                    return;
+                                }
+
                                 // Delete the file
                                 core.deleteFile(com, rowData.file_id);
